added all components together
